+++ R/data_plot.R
@@ -76,6 +76,7 @@
 #'   \item \code{bayestestR::rope()}
 #'   \item \code{estimate::estimate_contrasts()}
 #'   \item \code{parameters::model_parameters()}
+#'   \item \code{parameters::n_factors()}
 #'   \item \code{parameters::parameters_simulate()}
 #'   \item \code{performance::binned_residuals()}
 #'   \item \code{performance::check_collinearity()}